requirements.txt
fastapi==0.111.0
jose==1.0.0
passlib==1.7.4
pydantic==2.7.1
python-dotenv==1.0.1
python_jose==3.3.0
starlette==0.37.2
